M1: split receiver behind build tags so agent compiles on linux
First milestone of the UAT plan in docs/testing-strategy.md. The Go
agent's core (queue, uploader, commander, reconciler, etc.) is already
platform-portable; only agent/receiver pulls in macOS-only XPC via CGo
(#cgo LDFLAGS: -framework Foundation, #include <xpc/xpc.h>). That single
package was making `GOOS=linux go build ./agent/...` fail with "build
constraints exclude all Go files in agent/receiver", which in turn
blocks the headless-agent + server integration job (M3) and any future
Linux-side developer ergonomics.

This change:

- Tags the existing CGo files //go:build darwin (receiver.go, callbacks.go)
  so they are explicit about their platform, not implicit via the cgo
  cross-compile default of CGO_ENABLED=0.
- Extracts platform-neutral identifiers (Event struct, Error* constants,
  SetLogger + logger atomic) into a new untagged common.go so the linux
  build sees them without duplicating definitions.
- Adds receiver_other.go (//go:build !darwin) with a stub Receiver that
  mirrors the darwin Receiver's public surface (New / Events / Errors /
  Connect / SendApplicationControl / Ping / Disconnect). Connect,
  SendApplicationControl, and Ping return a sentinel ErrUnsupported so the
  agent fails closed on linux today; the inject-able variant for tests is
  M2's scope.
- Adds task build:agent:linux as the named acceptance check and wires it
  into the agent-test CI job (runs on macos-latest via cross-compile) so
  the platform-portability invariant stays green going forward.

Verified locally: task build:agent (darwin), task build:agent:linux, task
test:go:agent, task lint:go all pass. xpcbridge needs no changes because
it has no .go files; only the darwin receiver references its C headers.
hostid is already pure-Go and compiles on linux today.

No production darwin behaviour change. Refs UAT plan M1 milestone.
